+++ test/pedestal_toolbox/params_test.clj
@@ -102,3 +102,14 @@
                               {:date "4 score and 20 years ago"}}}
                             enter
                             (get-in [:response :body])))))))))
+
+(deftest query-param-accept-test
+  (let [params {"a" 1 :accept "application/csv"}
+        ctx {:request {:params params}}
+        enter (:enter query-param-accept)
+        ctx-with-updated-headers (enter ctx)]
+    (testing "a query param of content type is placed in the headers"
+      (is (= "application/csv"
+             (get-in ctx-with-updated-headers [:request :headers "accept"]))))
+    (testing "the original params aren't mutated"
+      (is (= (get-in ctx [:request :params "a"]) 1)))))